[Chore] Relocate ErrorPage into components
